fix: Ensure powersoftau export json uses stringify from field
