+++ src/r1cs_export_json.js
@@ -18,25 +18,7 @@
 */
 
 import {readR1cs}  from "r1csfile";
-
-export function stringifyBigInts(Fr, o) {
-    if (o instanceof Uint8Array)  {
-        return Fr.toString(o);
-    } else if (Array.isArray(o)) {
-        return o.map(stringifyBigInts.bind(null, Fr));
-    } else if (typeof o == "object") {
-        const res = {};
-        const keys = Object.keys(o);
-        keys.forEach( (k) => {
-            res[k] = stringifyBigInts(Fr, o[k]);
-        });
-        return res;
-    } else if ((typeof(o) == "bigint") || o.eq !== undefined)  {
-        return o.toString(10);
-    } else {
-        return o;
-    }
-}
+import { stringifyBigIntsWithField } from "./misc";
 
 
 export default async function r1csExportJson(r1csFileName, logger) {
@@ -46,5 +28,5 @@ export default async function r1csExportJson(r1csFileName, logger) {
     delete cir.curve;
     delete cir.F;
 
-    return stringifyBigInts(Fr, cir);
+    return stringifyBigIntsWithField(Fr, cir);
 }
